requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.2.10
attrs==20.2.0
autopep8==1.5.4
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.1.5
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.1
django-autoslug==1.9.8
django-baton==1.13.1
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-modal-forms==2.0.0
django-bootstrap4==2.3.1
django-braces==1.14.0
django-cors-headers==3.5.0
django-filter==2.4.0
django-geojson==3.1.0
django-heroku==0.3.1
django-jazzmin==2.4.1
django-leaflet==0.27.1
django-markdown2==0.3.1
django-markdownify==0.8.1
django-oauth-toolkit==1.3.3
django-pure-pagination==0.3.0
django-rest-authtoken==2.1.3
django-rest-framework==0.1.0
django-rest-framework-social-oauth2==1.1.0
django-templated-mail==1.1.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.1
djangorestframework-gis==0.16
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
docopt==0.6.2
dynaconf==3.1.2
entrypoints==0.3
filelock==3.0.12
firebase==3.0.1
firebase-admin==4.4.0
Flask==1.1.2
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.4
future==0.18.2
gcloud==0.18.3
gem==0.1.12
generics==2.0.0
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-cloud==0.34.0
google-cloud-core==1.4.3
google-cloud-firestore==1.9.0
google-cloud-storage==1.32.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
gTTS==2.1.1
gTTS-token==1.1.3
gunicorn==20.0.4
httplib2==0.18.1
hvac==0.10.5
idna==2.10
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jws==0.1.3
knox==0.1.11
loguru==0.5.3
Mako==1.1.3
Markdown==3.2.1
markdown2==2.3.9
markdownify==0.5.3
MarkupSafe==1.1.1
mdutils==1.3.0
mistune==0.8.4
movies==1.0.0
msgpack==1.0.0
mysqlclient==1.4.4
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.4
numpy==1.19.2
oauth2client==4.1.2
oauthlib==3.1.0
packaging==20.4
pandas==1.1.2
pandocfilters==1.4.2
parso==0.7.1
pg==0.1
pickleshare==0.7.5
Pillow==7.2.0
pip==20.3.1
pip-autoremove==0.9.1
pipdeptree==1.0.0
pipreqs==0.4.10
postgis==1.0.4
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.13.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.9.8
Pygments==2.6.1
PyJWT==1.7.1
PyMySQL==0.10.1
PyOpenGL==3.1.5
pyparsing==2.4.7
Pyrebase4==4.4.1
pyrsistent==0.17.0
python-application==2.8.0
python-dateutil==2.8.1
python-editor==1.0.4
python-jwt==2.0.1
python3-openid==3.2.0
pytz==2020.1
pywinpty==0.5.7
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.6
Send2Trash==1.5.0
setuptools==50.3.2
shortuuid==1.0.1
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.0.1
SQLAlchemy==1.2.13
sqlparse==0.3.1
static3==0.7.0
terminado==0.8.3
testpath==0.4.4
tkintertable==1.3.2
toml==0.10.1
tornado==6.0.4
traitlets==5.0.4
uritemplate==3.0.1
urllib3==1.25.10
validators==0.18.1
virtualenv==20.0.35
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wheel==0.35.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
win32-setctime==1.0.3
yarg==0.1.9
yarn==0.0.9
zope.interface==5.2.0
